requirements.txt
torch==2.4.1
numpy
pandas
matplotlib
tqdm
wandb
pyarrow
huggingface_hub
torchvision
fastapi
uvicorn
click
graphviz
python-dotenv
azure-storage-blob
